requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
attrs==20.3.0
black==20.8b1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dill==0.3.3
flake8==3.8.4
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.0
importlib-resources==3.3.0
iniconfig==1.1.1
joblib==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pathspec==0.8.1
Pillow==8.0.1
pluggy==0.13.1
promise==2.3
protobuf==3.14.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-datasets==4.1.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-metadata==0.25.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.54.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
